fix: make the links in /forms work in local env (#16)
Without the port number in the `Host` header, the /formList endpoint in aether-odk will not include the port number. This means that ODK Collect will receive invalid form URLs and will not be able to download any forms.

(Duplicate of https://github.com/eHealthAfrica/aether/pull/333, but how can we share conf files between the repos? 🤷‍♂️ )